+++ cli.h
@@ -4,5 +4,6 @@
 #include "common.h"
 
 gboolean cli_init(int *pargc, char ***pargv);
+extern gboolean autostart;
 
 #endif